# Assistant/mobile_client/openapi.json
{"openapi":"3.1.0","info":{"title":"Smart Health AI API","description":"Clinical Decision Support for Elderly Care","version":"0.3.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/cache-stats":{"get":{"summary":"Cache Stats","description":"Diagnostic — requires auth in production.","operationId":"cache_stats_cache_stats_get","parameters":[{"name":"X-API-Key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/stats":{"get":{"summary":"Rag Stats","description":"Diagnostic — requires auth in production.","operationId":"rag_stats_rag_stats_get","parameters":[{"name":"X-API-Key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/preview":{"get":{"summary":"Rag Preview","description":"Diagnostic: see what chunks RAG retrieves for a query. Requires auth.","operationId":"rag_preview_rag_preview_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","title":"Q"}},{"name":"k","in":"query","required":false,"schema":{"type":"integer","default":3,"title":"K"}},{"name":"X-API-Key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/analyze-vitals":{"post":{"summary":"Post Analyze Vitals","operationId":"post_analyze_vitals_analyze_vitals_post","parameters":[{"name":"X-API-Key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vitals"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyzeVitalsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat":{"post":{"summary":"Post Chat","operationId":"post_chat_chat_post","parameters":[{"name":"X-API-Key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat/stream":{"post":{"summary":"Post Chat Stream","description":"Server-Sent Events stream of the LLM response.\n\nProtocol:\n  event: start       -> (open the stream)\n  event: chunk       -> data: {\"text\": \"partial\"}\n  event: final       -> data: {\"full_answer\", \"emergency\", \"severity\", ...}\n\nMobile app consumes this via EventSource or a polyfill.","operationId":"post_chat_stream_chat_stream_post","parameters":[{"name":"X-API-Key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AnalyzeVitalsResponse":{"properties":{"severity":{"type":"string","title":"Severity"},"summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Summary"},"alerts":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Alerts"}},"type":"object","required":["severity","summary","alerts"],"title":"AnalyzeVitalsResponse"},"ChatMessage":{"properties":{"role":{"type":"string","enum":["user","assistant","system"],"title":"Role"},"content":{"type":"string","title":"Content"}},"type":"object","required":["role","content"],"title":"ChatMessage"},"ChatRequest":{"properties":{"question":{"type":"string","title":"Question"},"chat_history":{"anyOf":[{"items":{"$ref":"#/components/schemas/ChatMessage"},"type":"array"},{"type":"null"}],"title":"Chat History"},"vitals":{"anyOf":[{"$ref":"#/components/schemas/Vitals"},{"type":"null"}]},"patient":{"anyOf":[{"$ref":"#/components/schemas/Patient"},{"type":"null"}]},"activity":{"anyOf":[{"type":"string","enum":["sitting","walking","running","climbing_stairs","going_downstairs","sleeping","standing","lying"]},{"type":"null"}],"title":"Activity"},"recent_events":{"anyOf":[{"items":{"$ref":"#/components/schemas/HealthEvent"},"type":"array"},{"type":"null"}],"title":"Recent Events"},"user_role":{"anyOf":[{"type":"string","enum":["wearer","caregiver"]},{"type":"null"}],"title":"User Role"},"vitals_trend":{"anyOf":[{"type":"string","maxLength":300},{"type":"null"}],"title":"Vitals Trend","description":"Optional short trend summary, e.g. 'HR rose 70->95 over 2h'"}},"type":"object","required":["question"],"title":"ChatRequest"},"ChatResponse":{"properties":{"answer":{"type":"string","title":"Answer"},"refused":{"type":"boolean","title":"Refused"},"model":{"type":"string","title":"Model"},"severity":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Severity"},"vitals_summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Vitals Summary"},"emergency":{"type":"boolean","title":"Emergency","default":false},"emergency_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Emergency Reason"},"recommended_action":{"type":"string","enum":["call_911","contact_caregiver","monitor","none"],"title":"Recommended Action","default":"none"},"red_flags":{"items":{"$ref":"#/components/schemas/RedFlagOut"},"type":"array","title":"Red Flags","default":[]},"drug_warnings":{"items":{"$ref":"#/components/schemas/DrugWarningOut"},"type":"array","title":"Drug Warnings","default":[]},"sources":{"items":{"$ref":"#/components/schemas/SourceOut"},"type":"array","title":"Sources","default":[]},"follow_ups":{"items":{"type":"string"},"type":"array","title":"Follow Ups","default":[]},"structured":{"$ref":"#/components/schemas/StructuredOut","default":{"key_points":[],"recommendations":[]}},"usage":{"$ref":"#/components/schemas/UsageOut","default":{}},"from_cache":{"type":"boolean","title":"From Cache","default":false},"latency_ms":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Latency Ms"}},"type":"object","required":["answer","refused","model"],"title":"ChatResponse"},"DrugWarningOut":{"properties":{"level":{"type":"string","title":"Level"},"category":{"type":"string","title":"Category"},"drugs":{"items":{"type":"string"},"type":"array","title":"Drugs"},"message":{"type":"string","title":"Message"},"citation":{"type":"string","title":"Citation"}},"type":"object","required":["level","category","drugs","message","citation"],"title":"DrugWarningOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthEvent":{"properties":{"type":{"type":"string","enum":["fall","tachycardia","hypoxia","geofence_exit","sos","other"],"title":"Type"},"when":{"type":"string","title":"When","description":"Human-readable time, e.g. '3 min ago' or ISO timestamp"},"detail":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Detail"}},"type":"object","required":["type","when"],"title":"HealthEvent"},"Patient":{"properties":{"age":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Age"},"sex":{"anyOf":[{"type":"string","enum":["M","F"]},{"type":"null"}],"title":"Sex"},"conditions":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Conditions"},"medications":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Medications"}},"type":"object","title":"Patient"},"RedFlagOut":{"properties":{"category":{"type":"string","title":"Category"},"label":{"type":"string","title":"Label"},"severity":{"type":"string","title":"Severity"},"matched_text":{"type":"string","title":"Matched Text"}},"type":"object","required":["category","label","severity","matched_text"],"title":"RedFlagOut"},"SourceOut":{"properties":{"source":{"type":"string","title":"Source"},"chunk":{"type":"integer","title":"Chunk"},"snippet":{"type":"string","title":"Snippet"}},"type":"object","required":["source","chunk","snippet"],"title":"SourceOut","description":"One retrieved corpus chunk supporting the answer."},"StructuredOut":{"properties":{"overview":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Overview"},"key_points":{"items":{"type":"string"},"type":"array","title":"Key Points","default":[]},"recommendations":{"items":{"type":"string"},"type":"array","title":"Recommendations","default":[]},"emergency":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Emergency"}},"type":"object","title":"StructuredOut","description":"Optional parsed sections of the answer for structured rendering.\nEmpty if the LLM didn't produce recognizable headings."},"UsageOut":{"properties":{"prompt_tokens":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Prompt Tokens"},"completion_tokens":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Completion Tokens"},"total_tokens":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Total Tokens"}},"type":"object","title":"UsageOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"Vitals":{"properties":{"hr":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Hr","description":"Heart rate, bpm"},"spo2":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Spo2","description":"Blood oxygen saturation, %"},"rr":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Rr","description":"Respiratory rate, br/min (optional)"},"temp":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Temp","description":"Temperature, °C"}},"type":"object","title":"Vitals"}}}}